+++ internal/vault/client_test.go
